% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-list.R
\name{cat.dwd.ftp.url}
\alias{cat.dwd.ftp.url}
\title{Print all URLs of the FTP server}
\usage{
cat.dwd.ftp.url()
}
\value{
A list containing all URLs of the DWD FTP server grouped
  by their type of measurement. It contains six elements
  \itemize{
    \item{\emph{observations.germany} - Corresponding to all links
            to observation data throughout Germany}
    \item{\emph{observations.global} - Corresponding to all links
            to observation data throughout the globe}
    \item{\emph{grid.germany} - Corresponding to all links
            to the grid data spanning Germany}
    \item{\emph{grid.europe} - Corresponding to all links
            to the grid data spanning Europe}
    \item{\emph{derived.germany} - Corresponding to all links
            to derived data within Germany}
    \item{\emph{regional.averages.germany} - Corresponding to all
            links regional averages throughout Germany}
  }
}
\description{
Provides a list of all possible download folders for
  a manual treatment.
}
\details{
If you, instead, want to choose the download folder
  interactively, please see the \code{\link{get.dwd.ftp.url}}
  function.
}
\seealso{
Other dwd-urls: \code{\link{get.dwd.ftp.url}}
}
\author{
Philipp Mueller
}
\concept{dwd-urls}
