% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{conversion.climate}
\alias{conversion.climate}
\title{Downloads data of the DWD}
\usage{
conversion.climate(files.list, files.description.list,
  csv.export = FALSE, download.folder, prefix.file.name, quiet = FALSE)
}
\arguments{
\item{files.list}{Named list of character vectors containing
absolute paths pointing towards the .zip files downloaded from
the DWD FTP server. The names correspond to \emph{recent},
\emph{historical}, and \emph{diverse}.}

\item{files.description.list}{Named list of the corresponding
description files. One for each element in \strong{files.list}.}

\item{csv.export}{If TRUE creates an additional folder containing
.csv files with the individual station data. Using this the data
can be used outside of R too. Default = FALSE.}

\item{download.folder}{This folder will be used to unpack and
extract the .zip archives in. It does not have to be the same as
the one containing the downloaded content.}

\item{prefix.file.name}{String, which will prepend to all save
files.}

\item{quiet}{Whether or not to display the output generated when
downloading the content. Default = FALSE.}
}
\value{
invisible( TRUE )
}
\description{
Downloads daily weather data from observation
  stations in Germany and extracts minimum and maximum temperature
  as well as precipitation data.

Since this function is agnostic of the type of data
  set picked for download and extraction, a prefix to the file
  names must be provided using \strong{prefix.file.name}. Else,
  the temperatures of e.g. both the hourly and daily data will be
  saved in a file called dwd.temperatures and one overrides the
  other. Instead, it will be saved into
  \emph{dwd.[prefix.file.name].temperatures}.

  The .csv files will be stored in
  \emph{[download.folder]/csv/[prefix.file.name]} and a separate
  folder will be created for each climatological quantity.
}
\author{
Philipp Mueller
}
