% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert.date.integer}
\alias{convert.date.integer}
\title{Convert integer objects into Date}
\usage{
\method{convert.date}{integer}(input.data, origin = "1970-01-01")
}
\arguments{
\item{input.data}{An integer vector containing the input date.}

\item{origin}{Day the numerical representation of the date will be
zero. Per default, as used in the POSIX standard, it is the
string "1970-01-01".}
}
\value{
Vector of \emph{Date} objects of the same length as the
  input vector \code{input.data}.
}
\description{
An auxiliary function to convert the date format used
  in the files on the FTP server of the DWD into a format, which
  can be handled by the \pkg{lubridate} package.
}
\author{
Philipp Mueller
}
