% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download.content}
\alias{download.content}
\title{Download files from the FTP server of the DWD}
\usage{
download.content(url, download.folder = NULL, quiet = FALSE,
  debug = FALSE)
}
\arguments{
\item{url}{Either a character vector of one or more URLs to
folders on the FTP server of the DWD or the output of the
\code{\link{get.dwd.ftp.url}} function.}

\item{download.folder}{Manual override for the default download
folder of the package. If NULL, the default value will be used
instead. Default = NULL.}

\item{quiet}{Whether or not to verbose the download
procedure. Default = FALSE.}

\item{debug}{If TRUE is enables verbose messages of the individual
downloads. Default = FALSE.}
}
\value{
Returns a character vector containing all directories data
  was written to terminated by a '/'.
}
\description{
Downloads the content of the supplied URLs. It is
  intended to work with URLs pointing at folders on the FTP.
}
\details{
The main reason why to use this function over the
  \strong{wget}, which is used internally for the download, is
  that it is automatically updates your data.

  When downloading content for the first time, the function
  reproduces the directory tree on the FTP server of the DWD and
  puts the specified files in the corresponding
  subfolders. Whenever the function is invoked another time, it
  compares the content of the folders in the URLs with the one of
  the corresponding local folders. All deleted files will be
  removed, new files will be downloaded, and all the others will
  stay untouched, which saves a lot of time when updating your
  data base.

  The folder, which will contain all the source files downloaded
  from the FTP server of the DWD, will be set via \code{option(
  "dwd2r.download.path" )}. Per default it is set to the
  \emph{R/dwd_data} folder in your home. If you wish to change
  this default path, you have to override this option in your
  \strong{.Rprofile} configuration file in your home. In addition,
  the download folder can be supplied manually using the
  \code{download.folder} argument.
}
\author{
Philipp Mueller
}
