% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{dwd.download}
\alias{dwd.download}
\title{Downloads data of the DWD}
\usage{
dwd.download(save.downloads = TRUE, csv.export = FALSE, url = NULL,
  download.folder = NULL, batch.choices = NULL,
  time.series.format = c("xts", "data.frame"),
  use.geospatial.position.format = TRUE, quiet = FALSE,
  debug = FALSE)
}
\arguments{
\item{save.downloads}{If TRUE, the downloaded \emph{.zip} files
are stored in \code{download.folder}\emph{/downloads_dwd}. Else,
they will be deleted after the extraction. Default = TRUE.}

\item{csv.export}{If TRUE, the function creates an additional
folder containing .csv files with the individual station
data. Using them, the data can be accessed outside of R
too. Default = FALSE.}

\item{url}{A character vector containing one or more URLs pointing
to the different data folders on the FTP server of the DWD. If
this input argument is present, the interactive selection of the
data set will be skip.}

\item{download.folder}{Specifies where the data will be stored and
downloaded to. If you wish to alter the default settings and
provide a string using this input argument instead, it is
advised to store it in the \code{options( "dwd2r.download.path"
)} too, because it will also be used to import the saved
data. You can overwrite its default value of
\emph{"~/R/dwd_data/"} by adding \code{options(
dwd2r.download.path = "PATH" )} to the \emph{.Rprofile}
configuration file in your home.}

\item{batch.choices}{Numerical vector containing the numbers,
which correspond to the choices in the interactive mode. If
NULL, the choices will be selected interactively. Default =
NULL.}

\item{time.series.format}{Format of the extracted time
series. They can either be of type \strong{data.frame} and
contain two columns, "date" and "value", or a time series
provided by the \pkg{xts} package. Default = "xts".}

\item{use.geospatial.position.format}{If FALSE, the object
containing the geospatial information of all stations will be of
type \strong{data.frame} and consist of the columns named
\emph{longitude}, \emph{latitude}, \emph{altitude}, and
\emph{name}. If TRUE, an object of class
\code{\link[sp]{SpatialPointsDataFrame}} will be used instead
and the \emph{altitude} and \emph{name} information can be
accessed via the \code{@data} attribute. Default = TRUE.}

\item{quiet}{Whether or not to display the output generated when
downloading the content. Default = FALSE.}

\item{debug}{If TRUE is enables verbose messages of the individual
downloads performed within \code{download.content}. Default =
FALSE.}
}
\value{
\code{invisible( TRUE )}
}
\description{
It will guide the user through the hierarchical
  structure of the DWD's FTP server and let her choose some
  data. Then, the data will the downloaded, extracted locally, and
  converted into a format usable within R. This data, along with
  some meta-data obtained during the download as well, will be
  stored on the hard disk so the user load it later on using the
  \code{\link{source.data}} function.
}
\details{
Since there is A LOT of different data provided by
  the German weather service (DWD), it makes no sense to specify
  them all using input arguments to this function. Instead, the
  user will be guided interactively through the different data
  sources and specifies the data she want using the inputs at the
  prompt. If you prefer to run the function non-interactive, you
  can use the \code{batch.choices} argument and supply a vector of
  numbers corresponding to your choices. See the
  \code{\link{get.dwd.ftp.url}} function for details.

  When downloading content for the first time, the function
  reproduces the directory tree on the FTP server of the DWD and
  puts the specified files in the corresponding
  subfolders. Whenever the function is invoked another time, it
  compares the content of the folders in the URLs with the one of
  the corresponding local folders. All deleted files will be
  removed, new files will be downloaded, and all the others will
  stay untouched, which saves a lot of time when updating your
  data base. See \code{\link{download.content}} for details.

  Depending on the data format present in the downloaded files,
  more than one file will be stored on the hard disk of the
  user. In case of aggregated data, a separate file per column
  (climatological quantity) will be saved using the name of the
  quantity as filename. In all those files in \strong{.RData}
  format, two objects will be contained: A named list (using the
  station names) of all the station data in the format specified
  by \code{time.series.format} and an object containing additional
  meta-information for all stations (their longitude, latitude,
  and altitude) in a format specified by
  \code{use.geospatial.position.format}. See
  \code{\link{conversion.climate}} for details.
}
\author{
Philipp Mueller
}
