% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-retrieval.R
\name{dwd2r.url.check}
\alias{dwd2r.url.check}
\title{Checking the validity of URLs}
\usage{
dwd2r.url.check(url)
}
\arguments{
\item{url}{Either a simple string or a character vector of several
URLs.}
}
\value{
invisible( TRUE ) if all URLs exist. Throws an error if
  not.
}
\description{
This helper functions tests whether URLs are
  reachable and throws an error if this is not the case.
}
\author{
Philipp Mueller
}
