% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{extract.content.climate}
\alias{extract.content.climate}
\title{Unzips files and extracts the content of DWD source files
  for a single station.}
\usage{
extract.content.climate(station.id, files.list, download.folder.recent,
  download.folder.historical)
}
\arguments{
\item{station.id}{String specifying a station.}

\item{files.list}{Named list of character vectors containing
absolute paths pointing towards the .zip files downloaded from
the DWD FTP server. The names correspond to \emph{recent} and
\emph{historical}.}

\item{download.folder.recent}{Folder, which will contain the
extracted recent archive.}

\item{download.folder.historical}{Folder, which will contain the
extracted historical archive.}
}
\value{
Named list of \pkg{xts}-class objects where each
  element of the list corresponds to one data column in the data
  files.
}
\description{
Reads in the content of both the recent and
  historical files corresponding to a single station, combines
  its data, and converts it into a format R can handle.
}
\details{
The station is specified by its \emph{station.id}. The
  extraction will take place in either
  \strong{download.folder.recent} or
  \strong{download.folder.historical}. Before extraction all its
  content will be removed.
}
\author{
Philipp Mueller
}
