% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-retrieval.R
\name{get.dwd.ftp.url}
\alias{get.dwd.ftp.url}
\title{Obtain URLs to the FTP server}
\usage{
get.dwd.ftp.url()
}
\value{
List containing to two elements
  \itemize{
    \item{ \emph{data} - a character vector containing one or more
             URLs pointing to the different data folders. }
    \item{ \emph{meta} - character vector pointing to the folder
             containing the meta data. Or NULL if there is no
             distinct folder and the meta data are contained in
             the folders present in the \emph{data} URLs } 
  }
}
\description{
Interactive function to extract the download URLs of the
  FTP server of the DWD
}
\details{
Since there is A LOT of different data provided by
  the German weather service (DWD), it makes not sense to specify
  them all using input arguments to this function. Instead, the
  user will be guided interactively through the different data
  sources and specifies the data she want using the inputs at the
  prompt.

  As another option, you can use the \code{\link{cat.dwd.ftp.url}}
  function, which returns a vector containing all download
  URLs. This vector can then be manipulated using e.g. the grep
  command.

  Since there is a lot of different set of data, it is also not
  possible to provide documentation for the individual data
  sources. Please see the \url{ftp://ftp-cdc.dwd.de/pub/CDC/} FTP
  server for further details.

  Before using the data of the DWD, be sure you read its
  \url{ftp://ftp-cdc.dwd.de/pub/CDC/Terms_of_use.pdf}!
}
\seealso{
Other dwd-urls: \code{\link{cat.dwd.ftp.url}}
}
\author{
Philipp Mueller
}
