% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import.file.metadata.climate}
\alias{import.file.metadata.climate}
\title{Imports the metadata of a single data .zip folder into R}
\usage{
import.file.metadata.climate(file.path)
}
\arguments{
\item{file.path}{Path to a single Metadaten_Geographie_* file
contained in the zip archives of the DWD.}
}
\value{
A data.frame containing
  \itemize{
    \item{\emph{id} - The ID of the station}
    \item{\emph{name} - Its name}
    \item{\emph{longitude} - Its longitude in degree}
    \item{\emph{latitude} - Its latitude in degree}
    \item{\emph{altitude} - Its altitude in meters}
  }
}
\description{
If the actual measurement station was relocated, the
  DWD sticks to the same name and ID but adds another line in the
  Metadaten_Gepgraphie_ file. In this function I will only access
  the most recent position (contained in the last line).
}
\author{
Philipp Mueller
}
