% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{source.data}
\alias{source.data}
\title{Load a data file into R}
\usage{
source.data(download.folder = NULL, envir = NULL)
}
\arguments{
\item{download.folder}{Specifies the folder in which the function
will look for \emph{.RData} files recursively. Per default the
\emph{R/dwd_data/} directory in the home folder will be
used. You can overwrite this behavior by setting \code{options(
dwd2r.download.path = "PATH" )} in the \emph{.Rprofile}
configuration file in your home.}

\item{envir}{Environment the data will be attached to. If not
specified, the data will be loaded to the environment the
function is called from. Default = NULL.}
}
\value{
\code{invisible()} but attaches the content of the chosen
  \emph{.RData} file to the specified R environment.
}
\description{
Searches the \emph{~/R/dwd_data/} directory or a
  specified folder for \emph{.RData} files recursively and
  displays its findings so the user can choose one of them.
}
\details{
In order to use the data with the \pkg{climex} package,
  it should be of class \pkg{xts} or of lists of class \pkg{xts}
  objects.

  If you do not have any data yet, use the
  \code{\link{dwd.download}} function to get some.
}
\author{
Philipp Mueller
}
\concept{import}
