% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{conversion.climate}
\alias{conversion.climate}
\title{Converts the downloaded files into a format usable within R}
\usage{
conversion.climate(files.list, files.description.list,
  csv.export = FALSE, download.folder, prefix.file.name,
  time.series.format = c("xts", "data.frame"),
  use.geospatial.position.format = TRUE, quiet = FALSE)
}
\arguments{
\item{files.list}{Named list of character vectors containing
absolute paths pointing towards the .zip files downloaded from
the DWD FTP server. The names correspond to \emph{recent},
\emph{historical}, and \emph{diverse}.}

\item{files.description.list}{Named list of the corresponding
description files. One for each element in \code{files.list}.}

\item{csv.export}{If TRUE, the function creates an additional
folder containing .csv files with the individual station
data. Using them, the data can be accessed outside of R
too. Default = FALSE.}

\item{download.folder}{This folder will be used to unpack and
extract the \emph{.zip} archives in. It does not have to be the
same as the one containing the downloaded content.}

\item{prefix.file.name}{String, which will be prepended to all
saved files.}

\item{time.series.format}{Format of the extracted time
series. They can either be of type \strong{data.frame} and
contain two columns, "date" and "value", or a time series
provided by the \pkg{xts} package. Default = "xts".}

\item{use.geospatial.position.format}{If FALSE, the object
containing the geospatial information of all stations will be of
type \strong{data.frame} and consist of the columns named
\emph{longitude}, \emph{latitude}, \emph{altitude}, and
\emph{name}. If TRUE, an object of class
\code{\link[sp]{SpatialPointsDataFrame}} will be used instead
and the \emph{altitude} and \emph{name} information can be
accessed via the \code{@data} attribute. Default = TRUE.}

\item{quiet}{Whether or not to display the output generated when
downloading the content. Default = FALSE.}
}
\value{
\code{invisible( TRUE )}
}
\description{
It takes a list of strings pointing to the individual
  files downloaded from the DWD server, extracts their content,
  converts it into a format usable within R. In addition, it also
  extracts meta-information about the stations and
  stores all of it on the hard disk of the user.
}
\details{
Depending on the data format present in the downloaded
  files, this function will store more than one file on the hard
  disk of the user. In case of aggregated data, a separate file
  per column (climatological quantity) will be saved using the
  name of the quantity as filename. In all those files in
  \strong{.RData} format, two objects will be contained: A named
  list (using the station names) of all the station data in the
  format specified by \code{time.series.format} and an object
  containing additional meta-information for all stations (their
  longitude, latitude, and altitude) in a format specified by
  \code{use.geospatial.position.format}.

  Since this function is agnostic of the type of data
  set picked for download and extraction, a prefix to the file
  names must be provided using \code{prefix.file.name}. Else,
  the temperatures of, e.g., both the hourly and daily data will be
  saved in a file called dwd.temperatures and one overrides the
  other. Instead, it will be saved into
  \code{dwd.[prefix.file.name].temperatures}.

  If the user requires an additional storing of all content in
  .csv files, they will be stored in
  \code{[download.folder]/csv/[prefix.file.name]} and a separate
  folder will be created for each climatological quantity.
}
\author{
Philipp Mueller
}
