% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{dwd.download}
\alias{dwd.download}
\title{Downloads data of the DWD}
\usage{
## Downloading the aggregated collection of the daily measured
## climatic data of the DWD into a local folder, create .csv files
## from all the time series, and store the files (to be able to do
## a diff at a latter point in time).
dwd.download( save.downloads = TRUE, csv.export = TRUE,
              download.folder = "./local/folder",
              batch.choices  = c( 1, 1, 5, 1 ),
              time.series.format = "xts" )
}
\arguments{
\item{save.downloads}{If TRUE the downloaded .zip files are stored
in download.folder/downloads_dwd. Else they will be deleted
after the extracting. Default = TRUE.}

\item{csv.export}{If TRUE creates an additional folder containing
.csv files with the individual station data. Using this the data
can be used outside of R too. Default = FALSE.}

\item{url}{If the user wants to avoid the guided selection.}

\item{download.folder}{Specifies the data will be stored and
downloaded too. It is advised to store it in the path stored in
the \code{options( "dwd2r.download.path" )}, which is also used
for importing the saved data. You can overwrite its default
value of "~/R/dwd_data/" by adding \code{options(
dwd2r.download.path = "PATH" )} to your .Rprofile path in your
home.}

\item{batch.choices}{Numerical vector containing the numbers,
which corresponds to the choices in the interactive mode. If
NULL, the choices will be done interactively. Default = NULL.}

\item{time.series.format}{Format of the extracted time
series. They can either be of type \strong{data.frame} and contain two
columns, "date" and "value", or a time series provided by the
\pkg{xts} package. Default = "xts".}

\item{use.geospatial.position.format}{If FALSE, the object
containing the geospatial information of all stations will be of
type data.frame and consists of the columns named
\emph{longitude}, \emph{latitude}, \emph{altitude}, and
\emph{name}. If TRUE, an object of class SpatialPointsDataFrame
will be used instead and \emph{altitude} and \emph{name}
information can be accessed via the \code{@data}
attribute. Default = TRUE.}

\item{quiet}{Whether or not to display the output generated when
downloading the content. Default = FALSE.}

\item{debug}{If TRUE is enables verbose messages of the individual
downloads performed within \code{download.content}. Default =
FALSE.}
}
\value{
invisible( TRUE )
}
\description{
Downloads daily weather data from observation
  stations in Germany and extracts minimum and maximum temperature
  as well as precipitation data.
}
\details{
The download will be done using 'wget'. Per default the
  dwd2r.download.path variable from the \code{getOption(
  "dwd2r.download.path" )} will be used to set the download
  path. Since this function will check the files already present
  it's strongly recommended to use the save.downloads
  options. Whenever this function is invoked again only updated
  files will be downloaded which saves a lot of traffic and
  time. The csv.export option can be used to export the time
  series into a data type file making it available outside of R
  too. In addition, the geographic positions of the individual
  stations will be extracted and saved as well. They are needed
  for the leaflet module of the climex shiny app. CAUTION: since
  this procedure takes a while its run in parallel on all cores of
  your machine!
}
\author{
Philipp Mueller
}
