% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-retrieval.R
\name{dwd2r.url.check}
\alias{dwd2r.url.check}
\title{Checking the validity of URLs}
\usage{
dwd2r.url.check(url, number.of.tries = 5)
}
\arguments{
\item{url}{Either a simple string or a character vector of several
URLs.}

\item{number.of.tries}{Numerical value of how often the function
attempts to test each URL before reporting it to be
unavailable. Default = 5.}
}
\value{
\code{invisible( TRUE )} if all URLs exist. Throws an
  error if not.
}
\description{
This helper function tests whether URLs are reachable
  and throws an error if this is not the case.
}
\details{
The check will be performed \code{number.of.tries} per
  URL. The FTP server of the DWD is quite prone to identify the
  \code{\link[RCurl]{url.exists}} function as an DDOS attempt.
  Therefore, the function will wait some random time before
  attempting to check the URL. The waiting time will be normal
  distributed with a mean of 150ms and a standard deviation of
  50ms and the distribution is cut-off at 1ms to be the lowest
  possible value (of course we do not want to have negative
  waiting times and the Gaussian distribution has infinite
  support).
}
\author{
Philipp Mueller
}
