% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{extract.content.climate}
\alias{extract.content.climate}
\title{Unzips and extracts the content of DWD source files
  for a single station}
\usage{
extract.content.climate(station.id, files.list, download.folder.recent,
  download.folder.historical, time.series.format = c("xts",
  "data.frame"))
}
\arguments{
\item{station.id}{String specifying a station.}

\item{files.list}{Named list of character vectors containing
absolute paths pointing towards the .zip files downloaded from
the DWD FTP server. The names correspond to \emph{recent} and
\emph{historical}.}

\item{download.folder.recent}{Folder, which will contain the
extracted recent archive.}

\item{download.folder.historical}{Folder, which will contain the
extracted historical archive.}

\item{time.series.format}{Format of the extracted time
series. They can either be of type \strong{data.frame} and contain two
columns, "date" and "value", or a time series provided by the
\pkg{xts} package. Default = "xts".}
}
\value{
- Named list of either \pkg{xts}-class or \strong{data.frame}
  objects where each element of the list corresponds to one
  column in the data files.
 - NULL if no data at all could be extracted for the station.
}
\description{
Reads in the content of both the recent and
  historical files corresponding to a single station, combines
  its data, and converts it into a format R can handle.
}
\details{
The station is specified by its \code{station.id}. The
  extraction will take place in either
  \code{download.folder.recent} or
  \code{download.folder.historical}. Before extraction all its
  content will be removed.
}
\examples{
## The \\pkg{dwd2r} provides some mock data so its internal
  ## functions can be tested and developed without downloading
  ## any content from the DWD server. At the package's root
  ## directory call this function in the following way (be sure to
  ## create the two folders \\code{download.folder.recent} and
  ## \\code{download.folder.historical} beforehand).
  dwd2r:::extract.content.climate( "03987",
    list( historical = "./res/produkt_03987_potsdam_historical_mock.zip",
          recent = "./res/produkt_potsdam_recent_03987_mock.zip" ),
          download.folder.recent, download.folder.historical, "xts" )

}
\author{
Philipp Mueller
}
