% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import.file.content.climate}
\alias{import.file.content.climate}
\title{Imports the content of a single produkt file into R}
\usage{
import.file.content.climate(file.path)
}
\arguments{
\item{file.path}{Path to a single produkt_* file contained in the
zip archives of the DWD.}
}
\value{
A list of all time series contained in the produkt file
  converted to the \pkg{xts} class and named according to some
  convention of the DWD files or according to the names of the
  corresponding rows.
}
\description{
It does the actual conversion of the format used by
  the DWD to the \pkg{xts} class. This function is intend to be
  used for aggregated station data.
}
\author{
Philipp Mueller
}
